Add sell product functionality (#17)

+++ src/main/java/com/teamknp/hotel/repository/KeyStatusRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KeyStatusRepository extends JpaRepository<KeyStatus, Integer> {
+public interface KeyStatusRepository extends JpaRepository<KeyStatus, Long> {
 }
